Прокинет моки в пропсы
