+++ project/src/index.tsx
@@ -1,6 +1,7 @@
 import App from './components/app/app';
 import React from 'react';
 import ReactDOM from 'react-dom';
+import { offers } from './mocks/mocks';
 
 const Setting = {
   PLACES_COUNT: 1000,
@@ -8,7 +9,7 @@ const Setting = {
 
 ReactDOM.render(
   <React.StrictMode>
-    <App placesCount={Setting.PLACES_COUNT} />
+    <App placesCount={Setting.PLACES_COUNT} offers={offers} />
   </React.StrictMode>,
   document.getElementById('root'),
 );
